backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
ultralytics==8.0.200
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy<2.0.03
requests==2.31.0
ollama>=0.0.11
google-generativeai>=0.3.0
inference-sdk>=0.9.0
